# smsc/xpmem: Refactor reg-cache to use tree's find() instead of iterate()