Hi, I'd like to propose some changes to the registration cache in smsc/xpmem. Currently, looking up a registration is an operation that iterates the whole tree (or until found). This is necessary as the tree is used to store attachments to different procs/endpoints, but can get slow and scale poorly. To address this, in this PR I suggest creating one regcache/tree per each endpoint/peer and using the `find()` method (O(logn) instead of O(n)).

This PR supersedes #11130.
FYI @devreal 

- The first commit implements the juicy part of this PR (find vs iterate).
- The second and smaller commit takes the opportunity to add support for the `MCA_RCACHE_FLAGS_PERSIST` and `MCA_RCACHE_FLAGS_CACHE_BYPASS` flags
    - Would this be better in its own PR instead of bundled here? If yes I can branch it out to a new one

Note that to put this together I took "bound" to mean the last valid byte in the range, inclusive, and fixed some instances where this was mishandled. Please let me know if this is not accurate/desired.

---

The benefit with this change will vary depending on the number of standing registrations and message size. I expect the relative gain to be relevant for small-to-medium messages (but larger than any copy-in-copy-out-to-single-copy thresholds). In a (single) node with 64 ranks, using mirobenchmarks, tuned's alltoall @ 1K improves from `~143 us to ~124 us`. Using my component (xhc), the hierarchical allreduce @ 16K improves from `~39 us to ~31 us`, while the (generally worse) non-hierarchical allreduce (more attachments/regs) @ 4K improves from `100 us to 42 us` .

While using a single tree for each endpoint will result in more trees, the total number of tree nodes does not increase.
A planned future enhancement that would cache peer-process-to-rcache information will further improve this. 

---

Signed-off-by: George Katevenis <gkatev@ics.forth.gr>