@gkatev 2023-02-09
@bosilca, below are the graphs from some performance measurements I had done, with the old iterate() approach versus the new find() one. They are from the OSU suite, on a node with 2x Epyc 7501 (64 procs total). In most cases the difference will be noticeable in relatively small sizes (but above the single-copy threshold), which is why I show up until 8 KB. The gap is also expected to widen in pontially busier non-microbenchmark scenarios with more active registrations.

<details>

![tuned_allgather](https://user-images.githubusercontent.com/31203804/217867348-30e2230b-673b-4fe8-8b4e-26f3fa653f54.svg)
![tuned_allreduce](https://user-images.githubusercontent.com/31203804/217867398-47d96ac0-1c73-4c03-823f-76ce214abd84.svg)
![tuned_alltoall](https://user-images.githubusercontent.com/31203804/217867422-04886772-f77a-45cc-b36e-2d2c8b9be2c8.svg)
![tuned_gather](https://user-images.githubusercontent.com/31203804/217867450-c349fd62-bd13-49dc-b7ac-52f294758e34.svg)
![tuned_reduce](https://user-images.githubusercontent.com/31203804/217867469-1bd989be-443c-4b89-bfe4-0de40007ed0a.svg)
![xhc_allreduce](https://user-images.githubusercontent.com/31203804/217867472-c9b855a0-bce7-4c91-b8e5-8eee12729a12.svg)
![xhc_flat_allreduce](https://user-images.githubusercontent.com/31203804/217867481-9ba696e8-e7a9-4dcb-962e-b252db489408.svg)

</details>